changed imports/exports to index.js structure and refactored code
