+++ src/components/atoms/DropdownMenu.jsx
@@ -1,10 +1,8 @@
 import React, { Component } from 'react'
 import styled from 'styled-components'
-import NavLink from './NavLink'
+import { NavLink } from '../atoms'
 
 /** Overview of DropdownMenu component:
- *  - Created a styled div container that wraps DropdownMenu
- *  - State is initially false
  *  - Have two functions that either display or hide the dropdown menu depending on mouse hover
  *  - DropdownMenu renders links passed down from NavButtons (Help and Signin)
  */
