+++ src/components/atoms/HomePageButton.jsx
@@ -1,10 +1,6 @@
 import React, { Component } from 'react'
 import Styled from 'styled-components'
 import { Link } from 'gatsby'
-/** Overview of index.js component:
- *  - Created a styled div container which will wrap our NavBar component and, later on, our app.
- *
- */
 
 const Container = Styled.div`
 #hero-image-button{
