+++ src/components/atoms/ProductCard.jsx
@@ -1,5 +1,3 @@
-// Component for rendering individual product cards
-
 import React, { Component } from 'react'
 import styled from 'styled-components'
 
@@ -35,7 +33,6 @@ const Container = styled.div`
   }
 `
 
-// Renders product information and first image from connected contentful data
 class ProductCard extends Component {
   render() {
     const product = this.props.product
@@ -46,6 +43,7 @@ class ProductCard extends Component {
           <img
             className="product-card-image"
             src={product.images[0].file.url}
+            alt="earring"
           />
           <p className="product-card-name">{product.title}</p>
           <p className="product-card-price">${product.price}</p>
