+++ src/components/atoms/ShoppingBagIcon.jsx
@@ -27,7 +27,6 @@ const Container = styled.div`
 
 class ShoppingBagIcon extends Component {
   render() {
-    const { click } = this.props
     return (
       <Container>
         <img
