+++ src/components/atoms/TryptychPanel.jsx
@@ -1,5 +1,3 @@
-// Component for each panel on tryptych, rendering passed in image and information
-
 import React, { Component } from 'react'
 
 class TryptychPanel extends Component {
