+++ src/components/atoms/index.jsx
@@ -2,3 +2,11 @@ export { default as DropdownMenu } from './DropdownMenu'
 export { default as NavLink } from './NavLink'
 export { default as ShoppingBagIcon } from './ShoppingBagIcon'
 export { default as SideNavLink } from './SideNavLink'
+export { default as ProductCard } from './ProductCard'
+export {
+  default as ProductPageProductDetails,
+} from './ProductPageProductDetails'
+export { default as ProductPageProductInfo } from './ProductPageProductInfo'
+export { default as ProductPageProductOrder } from './ProductPageProductOrder'
+export { default as TryptychPanel } from './TryptychPanel'
+export { default as HomePageButton } from './HomePageButton'
