+++ src/components/index.js
@@ -1,15 +1,10 @@
-export {default as Header } from './Header'
-export {default as HomePageHeroImage } from './HomePageHeroImage'
-export {default as HomePageTryptych } from './HomePageTryptych'
-export {default as Image } from './Image'
-export {default as Layout } from './Layout'
-export {default as NavBar } from './NavBar'
-export {default as LayoutFont } from './LayoutFont'
-export {default as ProductCard } from './atoms/ProductCard'
-export {default as ProductList } from './molecules/ProductList'
-export {default as ProductPageContainer } from './ProductPageContainer'
-export {default as ProductPageInfoContainer } from './ProductPageInfoContainer'
-export {default as ProductPageProductDetails } from './atoms/ProductPageProductDetails'
-export {default as ProductPageProductOrder } from './ProductPageProductOrder'
-export {default as SideNavContainer } from './SideNavContainer'
-export {default as TryptychPanel } from './atoms/TryptychPanel'
+export { default as Header } from './Header'
+export { default as HomePageHeroImage } from './HomePageHeroImage'
+export { default as HomePageTryptych } from './HomePageTryptych'
+export { default as Image } from './Image'
+export { default as Layout } from './Layout'
+export { default as NavBar } from './NavBar'
+export { default as LayoutFont } from './LayoutFont'
+export { default as ProductPageContainer } from './ProductPageContainer'
+export { default as ProductPageInfoContainer } from './ProductPageInfoContainer'
+export { default as SideNavContainer } from './SideNavContainer'
