--- src/components/molecules/BannerPromo.css
@@ -1,4 +1,0 @@
-.banner-text {
-  font-style: italic;
-  font-weight: bold;
-}
