+++ src/components/molecules/index.jsx
@@ -1,3 +1,4 @@
 export { default as BannerPromo } from './BannerPromo'
 export { default as NavButtons } from './NavButtons'
 export { default as SideNav } from './SideNav'
+export { default as ProductList } from './ProductList'
