+++ src/pages/testProductPage.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { Link } from 'gatsby'
 import NavBar from '../components/NavBar'
 
 import Layout from '../components/Layout'
@@ -7,7 +6,7 @@ import ProductPageInfoContainer from '../components/ProductPageInfoContainer'
 
 const testProductPage = () => (
   <Layout>
-    <NavBar/>
+    <NavBar />
     <ProductPageInfoContainer />
   </Layout>
 )
